requirements.txt
appdirs==1.4.3
bleach==2.0.0
colorama==0.3.9
decorator==4.0.11
entrypoints==0.2.2
html5lib==0.999999999
inputs==0.1
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.0.1
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.0rc1
olefile==0.44
opencv-python==3.2.0.7
packaging==16.8
pandocfilters==1.4.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pyzmq==16.0.2
simplegeneric==0.8.1
six==1.10.0
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0